Functionality for PDFs and classes
I've put all the functions into classes in separate files as described in my UML for the last assignment. Except for getReply(). There is a problem when putting it in a class which i will fix soon. Also, it can now read and summarize PDFs. The CSS code has not been moved to a separate file yet.